First pass at Telnet communications with NLZ.
Implemented connecting to NLZ, disconnecting, logging on (albeit
manually, where you have to type your name and password in and hit
Enter to continue, because the user options aren't connnected to a
data store yet). You can also hit Enter to advance and read and/or
enter NLZ commands and display responses.