Implemented the word wrapping at 72 option.
If you enable word wrapping at 72 characters, then BIXpy wraps long
lines and their continuations at 72 characters, with the word break
character as the last character on the line (which can be the 73rd
position) until you enter a line that doesn't result in yet another
word wrap or you hit Enter by itself.

If you want a paragraph break, you'll have to hit Enter by itself
again. You do not need to enter a paragraph break if you want to
end the message, but you do have to end the word wrapping before
you can enter a . on a line by itself (which triggers the end of
text entry in NLZ).