Window sizing and memorizing enhancements.
Allow the text entry field to grow wider or narrower as the main
window is resized. Allow the text display area to grow or shrink
in both dimensions as the main window is resized. If the window
state is normal when the program exits, save both the position
and the size of the main window.