Use explicit utf-8 conversions when talking to NLZ
This way, there's no need to strip out characters
higher than '~' (as was the case when using ascii)
and unknown chars are replaced with a question mark
(or in some cases the original byte value as a hex
string, such as the pilcrow sign, which is B6 in
CP1252 and should be translated to U+00B6, which
is C2 B6 in utf-8, but shows up as 'B6' instead).

Also changed a byte array variable named text to
bytes, to avoid confusion regarding its contents.